+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - If POST request body has already been used, an error is thrown instead of returning an empty object
 
+### Added
+
+- Finally, `event` is no longer needed in the load function call to `superValidate`.
+
 ## [0.6.9] - 2023-04-06
 
 ### Fixed